Updates for load balancers
Before, we weren't getting anything but a reference to IPConfigurations
and PublicIPs. This fixes that, but there seems to be an issue with
finding Backend NIC IPs that causes a lot of errors to hit the screen. I
think it is still worth pushing this, but I need to investigate those
errors.